requirements.txt
img2pdf==0.3.6
numpy==1.19.0
opencv-contrib-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
tqdm==4.47.0
